Consolidated Interfaces for models in types/models directory
