+++ node-service/models/Scan.ts
@@ -1,78 +1,14 @@
 import mongoose, { Document, Schema, Types } from 'mongoose';
 
-export interface IVulnerability {
-  id: string;
-  severity: 'critical' | 'high' | 'medium' | 'low' | 'info';
-  description: string;
-  references?: string[];
-  fixedIn?: string;
-}
+import { IVulnerability } from '../types/models';
 
-export interface IDependency {
-  packageName: string;
-  currentVersion: string;
-  latestVersion?: string;
-  isOutdated: boolean;
-  vulnerabilities?: IVulnerability[];
-  filePath?: string;
-  dependencyType?: string;
-}
+import { IDependency } from '../types/models';
 
-// Define all possible scan states
-export type ScanState =
-  | 'initiated'
-  | 'dependency-scanning'
-  | 'dependencies-scanned'
-  | 'vulnerability-scanning'
-  | 'vulnerabilities-scanned'
-  | 'pr-creation'
-  | 'pr-created'
-  | 'completed'
-  | 'failed';
+import { ScanState } from '../types/models';
 
-// State transition history record
-export interface StateTransition {
-  state: ScanState;
-  timestamp: Date;
-  metadata?: Record<string, any>;
-}
+import { StateTransition } from '../types/models';
 
-export interface IScan extends Document {
-  _id: Types.ObjectId;
-  repositoryId: Types.ObjectId;
-  userId: Types.ObjectId;
-
-  // Traditional status fields
-  status: 'pending' | 'in-progress' | 'completed' | 'failed';
-
-  // New state machine fields
-  state: ScanState;
-  stateHistory: StateTransition[];
-
-  // Options
-  includeVulnerabilities: boolean;
-  createPR: boolean;
-
-  branch?: string;
-  commit?: string;
-  dependencies: IDependency[];
-  outdatedCount: number;
-  vulnerabilityCount: number;
-  highSeverityCount: number;
-
-  // PR information
-  prNumber?: number;
-  prUrl?: string;
-
-  // Error information
-  errorMessage?: string;
-
-  // Timestamps
-  startedAt?: Date;
-  completedAt?: Date;
-  createdAt: Date;
-  updatedAt: Date;
-}
+import { IScan } from '../types/models';
 
 const VulnerabilitySchema = new Schema<IVulnerability>({
   id: { type: String, required: true },
